requirements.txt
asgiref==3.7.2
Authlib==1.3.0
backports.zoneinfo==0.2.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.3
Django==4.2.10
django-oidc-provider==0.8.2
djangorestframework==3.14.0
drf-oidc-auth==3.0.0
exceptiongroup==1.2.0
future==0.18.3
idna==3.6
iniconfig==2.0.0
packaging==23.2
pluggy==1.4.0
pycparser==2.21
pycryptodomex==3.20.0
pyjwkest==1.4.2
pytest==8.0.1
pytest-django==4.8.0
pytz==2024.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
typing-extensions==4.9.0
urllib3==2.2.0
